+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+ - A mechanism for deriving: `datatype99` now can be specified with `derive(...)`.
+ - A deriver `dummy` which generates nothing.
+
 ## [1.0.0] - 2021-03-28
 
 ### Added